Reorganize stream buffer files
