+++ Source/FFmpegReader.h
@@ -28,20 +28,22 @@ namespace FFmpegInteropX
 {
     using namespace Concurrency;
 
-	ref class FFmpegReader
-	{
+    ref class StreamBuffer;
+
+    ref class FFmpegReader
+    {
     internal:
         FFmpegReader(AVFormatContext* avFormatCtx, std::vector<MediaSampleProvider^>* sampleProviders, MediaSourceConfig^ config);
 
         int ReadPacket();
         int ReadPacketForStream(StreamBuffer^ buffer);
-		void Start();
-		void Stop();
+        void Start();
+        void Stop();
         void Flush();
         HRESULT Seek(TimeSpan position, TimeSpan& actualPosition, TimeSpan currentPosition, bool allowFastSeek, MediaSampleProvider^ videoStream, MediaSampleProvider^ audioStream);
 
 
-	private:
+    private:
 
         ~FFmpegReader();
         bool TrySeekBuffered(TimeSpan position, TimeSpan& actualPosition, bool fastSeek, MediaSampleProvider^ videoStream, MediaSampleProvider^ audioStream);
@@ -49,8 +51,8 @@ namespace FFmpegInteropX
         void ReadDataLoop();
         void FlushCodecs();
 
-		AVFormatContext* avFormatCtx;
-		std::vector<MediaSampleProvider^>* sampleProviders;
+        AVFormatContext* avFormatCtx;
+        std::vector<MediaSampleProvider^>* sampleProviders;
         MediaSourceConfig^ config;
 
         std::mutex mutex;
@@ -63,210 +65,5 @@ namespace FFmpegInteropX
         bool isLastSeekForward;
         TimeSpan lastSeekStart;
         TimeSpan lastSeekActual;
-    };
-
-    ref class StreamBuffer
-    {
-    internal:
-        StreamBuffer(int streamIndex, MediaSourceConfig^ config)
-            : config(config)
-        {
-            StreamIndex = streamIndex;
-            
-        }
-
-        property int StreamIndex;
-
-        void QueuePacket(AVPacket* packet)
-        {
-            std::lock_guard<std::mutex> lock(mutex);
-            buffer.push_back(packet);
-            bufferSize += packet->size;
-        }
-
-        bool ReadUntilNotEmpty(FFmpegReader^ reader)
-        {
-            while (IsEmpty())
-            {
-                if (reader->ReadPacketForStream(this) < 0)
-                {
-                    DebugMessage(L"GetNextPacket reaching EOF\n");
-                    break;
-                }
-            }
-            return !IsEmpty();
-        }
-
-        bool SkipUntilTimestamp(FFmpegReader^ reader, LONGLONG target)
-        {
-            bool foundPacket = false;
-
-            while (!foundPacket)
-            {
-                if (ReadUntilNotEmpty(reader))
-                {
-                    // peek next packet and check pts value
-                    auto packet = PeekPacket();
-
-                    auto pts = packet->pts != AV_NOPTS_VALUE ? packet->pts : packet->dts;
-                    if (pts != AV_NOPTS_VALUE && packet->duration != AV_NOPTS_VALUE)
-                    {
-                        auto packetEnd = pts + packet->duration;
-                        if (packet->duration > 0 ? packetEnd <= target : packetEnd < target)
-                        {
-                            DropPackets(1);
-                        }
-                        else
-                        {
-                            foundPacket = true;
-                            break;
-                        }
-                    }
-                    else
-                    {
-                        break;
-                    }
-                }
-                else
-                {
-                    // no more packet found
-                    break;
-                }
-            }
-
-            return foundPacket;
-        }
-
-        bool IsEmpty()
-        {
-            std::lock_guard<std::mutex> lock(mutex);
-            return buffer.empty();
-        }
-
-        bool IsFull(MediaSampleProvider^ sampleProvider)
-        {
-            std::lock_guard<std::mutex> lock(mutex);
-            auto maxSize = config->ReadAheadBufferSize;
-            auto maxDuration = config->ReadAheadBufferDuration;
-
-            bool full = maxSize >= 0 && (long long)bufferSize > maxSize;
-            if (!full && maxDuration.Duration >= 0 && buffer.size() > 1)
-            {
-                auto firstPacket = buffer.front();
-                auto lastPacket = buffer.back();
-                auto firstPts = firstPacket->pts != AV_NOPTS_VALUE ? firstPacket->pts : firstPacket->dts;
-                auto lastPts = lastPacket->pts != AV_NOPTS_VALUE ? lastPacket->pts : lastPacket->dts;
-
-                if (firstPts != AV_NOPTS_VALUE && lastPts != AV_NOPTS_VALUE)
-                {
-                    auto duration = sampleProvider->ConvertDuration(lastPts - firstPts);
-                    full = duration > maxDuration;
-                }
-            }
-            return full;
-        }
-
-        AVPacket* PopPacket()
-        {
-            std::lock_guard<std::mutex> lock(mutex);
-
-            AVPacket* packet = NULL;
-            if (!buffer.empty())
-            {
-                packet = buffer.front();
-                buffer.pop_front();
-                bufferSize -= packet->size;
-            }
-
-            return packet;
-        }
-
-        AVPacket* PeekPacket()
-        {
-            std::lock_guard<std::mutex> lock(mutex);
-
-            AVPacket* packet = NULL;
-            if (!buffer.empty())
-            {
-                packet = buffer.front();
-            }
-
-            return packet;
-        }
-
-        AVPacket* PeekPacketIndex(int index)
-        {
-            std::lock_guard<std::mutex> lock(mutex);
-            return buffer.at(index);
-        }
-
-        int TryFindPacketIndex(LONGLONG pts, bool requireKeyFrame)
-        {
-            std::lock_guard<std::mutex> lock(mutex);
-
-            bool hasEnd = false;
-            int index = 0;
-            int result = -1;
-            for (auto packet : buffer)
-            {
-                if (packet->pts <= pts)
-                {
-                    if (!requireKeyFrame || packet->flags & AV_PKT_FLAG_KEY)
-                    {
-                        result = index;
-                        if (packet->pts + packet->duration >= pts)
-                        {
-                            hasEnd = true;
-                            break;
-                        }
-                    }
-                }
-                else
-                {
-                    hasEnd = true;
-                    break;
-                }
-                index++;
-            }
-
-            if (result >= 0 && !hasEnd)
-            {
-                result = -1;
-            }
-
-            return result;
-        }
-
-        void Flush()
-        {
-            std::lock_guard<std::mutex> lock(mutex);
-            while (!buffer.empty())
-            {
-                auto packet = buffer.front();
-                bufferSize -= packet->size;
-                buffer.pop_front();
-
-                av_packet_free(&packet);
-            }
-        }
-
-        void DropPackets(int count)
-        {
-            std::lock_guard<std::mutex> lock(mutex);
-            for (int i = 0; i < count; i++)
-            {
-                auto packet = buffer.front();
-                bufferSize -= packet->size;
-                buffer.pop_front();
-
-                av_packet_free(&packet);
-            }
-        }
-
-    private:
-        std::deque<AVPacket*> buffer;
-        std::mutex mutex;
-        size_t bufferSize;
-        MediaSourceConfig^ config;
     };
 }
